requirements.txt
discord.py==1.7.3
discord-components==2.0.4
python-dotenv==0.19.0
requests==2.26.0
chat-exporter==1.7
pymongo==4.2.0